@@ -47,7 +47,8 @@ public class ListenConsole extends Thread {
             String readLine = bufferedReader.readLine();
 
             if (peerList.size() == 0) {
-                System.out.println("no peers...");
+                System.out.println("no peers..., set log level to 400");
+                Log.LEVEL = 400;
                 continue;
             }
 
@@ -115,17 +116,6 @@ public class ListenConsole extends Thread {
 //                    System.out.println("Services last run: ConnectionHandler: " + (System.currentTimeMillis() - ConnectionHandler.lastRun) + " MessageDownloader: " + (System.currentTimeMillis() - MessageDownloader.lastRun) + " MessageVerifierHsqlDb: " + (System.currentTimeMillis() - MessageVerifierHsqlDb.lastRun));
 //                    System.out.println("Livetime socketio connections: " + Stats.getSocketioConnectionsLiveTime());
 
-                    int size = 0;
-                    ConnectionReaderThread.threadLock.lock();
-                    try {
-                        size = ConnectionReaderThread.threads.size();
-                    } finally {
-                        ConnectionReaderThread.threadLock.unlock();
-                    }
-
-
-                    System.out.println("Threads: " + size);
-
                     Map<String, List<DefaultPooledObjectInfo>> stringListMap = ByteBufferPool.getPool().listAllObjects();
 
                     String out = "";
@@ -143,7 +133,6 @@ public class ListenConsole extends Thread {
 
                     System.out.println("NodeStore blacklist: ");
                     serverContext.getNodeStore().printBlacklist();
-
 
                 } finally {
                     peerList.getReadWriteLock().writeLock().unlock();
@@ -183,6 +172,9 @@ public class ListenConsole extends Thread {
                 serverContext.getNodeStore().saveToDisk();
                 Server.shutdown(serverContext);
                 System.exit(0);
+            } else if (readLine.equals("cg")) {
+                serverContext.getNodeStore().clearGraph();
+                System.out.println("cleared node graph");
             } else if (readLine.equals("c")) {
                 System.out.println("closing all connections...");
 
