@@ -8,7 +8,7 @@ public class ServerContext {
     private int port;
     private LocalSettings localSettings;
     private final KadStoreManager kadStoreManager = new KadStoreManager(this);
-    private PeerList peerList = new PeerList();
+    private PeerList peerList = new PeerList(this);
     private NodeStore nodeStore;
     private NodeId nodeId;
     private KademliaId nonce;
