@@ -7,11 +7,12 @@ public class NodeStoreMaintainJob extends Job {
 
 
     public NodeStoreMaintainJob(ServerContext serverContext) {
-        super(serverContext, 1000L * 30L * 1L, true);
+        super(serverContext, 1000L * 5L * 1L, true);
     }
 
     @Override
     public void init() {
+        // no need for job setup
     }
 
     @Override
